Refactor parser object
