+++ .gitignore
@@ -29,3 +29,4 @@ _testmain.go
 *.exe
 *.test
 *.prof
+*.DS_Store
